+++ package.json
@@ -3,7 +3,8 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@ag-grid-enterprise/server-side-row-model": "^26.2.1",
+    "@ag-grid-community/react": "^27.3.0",
+    "@ag-grid-enterprise/server-side-row-model": "^27.3.0",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@material-ui/core": "^4.12.4",
@@ -13,11 +14,13 @@
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "ag-grid-community": "^26.2.1",
-    "ag-grid-enterprise": "^26.2.1",
-    "ag-grid-react": "^26.2.0",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
+    "@types/react": "^17.0.45",
+    "@types/react-dom": "^18.0.4",
+    "ag-grid-community": "^27.3.0",
+    "ag-grid-enterprise": "^27.3.0",
     "axios": "^0.24.0",
-    "bootstrap": "^5.1.3",
     "date-fns": "^2.28.0",
     "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
@@ -29,6 +32,7 @@
     "react-hook-form-mui": "^3.2.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
+    "typescript": "^4.6.4",
     "web-vitals": "^2.1.2"
   },
   "scripts": {
