+++ src/App.tsx
@@ -1,6 +1,5 @@
 import React, { useState, useEffect } from "react";
 import { Routes, Route, useNavigate } from "react-router-dom";
-// import "bootstrap/dist/css/bootstrap.min.css";
 import "./App.css";
 
 import AuthService from "./services/auth.service";
@@ -11,7 +10,6 @@ import CompanyList from "./components/CompanyList/CompanyList";
 import AddCompany from "./components/CompanyList/AddCompany";
 import EditCompany from "./components/CompanyList/EditCompany";
 
-// import AuthVerify from "./common/AuthVerify";
 import EventBus from "./common/EventBus";
 import ContactList from "./components/ContactList/ContactList";
 import AddContact from "./components/ContactList/AddContact";
