+++ src/common/EventBus.tsx
@@ -2,10 +2,10 @@ const eventBus = {
   on(event, callback) {
     document.addEventListener(event, (e) => callback(e.detail));
   },
-  dispatch(event, data) {
+  dispatch(event, data?) {
     document.dispatchEvent(new CustomEvent(event, { detail: data }));
   },
-  remove(event, callback) {
+  remove(event, callback?) {
     document.removeEventListener(event, callback);
   },
 };
