+++ src/components/ContactList/AddContact.tsx
@@ -1,43 +1,76 @@
-import { useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-import { Button, Stack } from "@mui/material";
-import { useForm } from "react-hook-form";
+import React, { useEffect, useState } from "react";
 import {
   FormContainer,
   TextFieldElement,
   SelectElement,
 } from "react-hook-form-mui";
+import { useNavigate } from "react-router-dom";
+import { Button, Stack, Alert } from "@mui/material";
+import { useForm } from "react-hook-form";
 
+import axios from "../../utils/config.util";
 import AuthService from "../../services/auth.service";
-import CompanyService from "../../services/company.service";
-import toast from "../../utils/toast.util";
 
-const AddCompany = () => {
+const API_URL = process.env.REACT_APP_API_ENDPOINT;
+
+const AddContact = () => {
   const navigate = useNavigate();
+  const [showError, setShowError] = useState("");
+  const [companies, setCompanies] = useState([]);
+  const [showSuccess, setShowSuccess] = useState("");
+
   const formContext = useForm({
     defaultValues: {},
   });
   const { reset } = formContext;
-  const handleClearForm = () => {
-    reset();
-  };
+  const handleClearForm = () => reset();
 
   const handleSubmitForm = (e) => {
-    if (!e.email || !e.name) return;
     const payload = { ...e };
-    CompanyService.create(payload)
+    axios
+      .post(API_URL + "contact/create", payload)
+      .then(({ data }) => data)
       .then((response) => {
+        console.log(response);
         if (response.status) {
-          toast.success(response.message);
+          setShowSuccess(response.message);
           reset();
         } else {
-          toast.error(response.message);
+          setShowError(response.message);
         }
       })
-      .catch(({ response }) => {
-        toast.error(response.message);
+      .catch((error) => {
+        console.log(error);
+        setShowError(error.response);
       });
   };
+
+  const statusList = [
+    {
+      id: "",
+      title: "Select",
+    },
+    {
+      id: "active",
+      title: "Active",
+    },
+    {
+      id: "inactive",
+      title: "Inactive",
+    },
+  ];
+
+  useEffect(() => {
+    axios
+      .get(API_URL + "company/listOfNames")
+      .then(({ data }) => data)
+      .then(({ result }) => {
+        setCompanies(result);
+      })
+      .catch((error) => {
+        setCompanies([]);
+      });
+  }, []);
 
   // check if user is authenticated, if not redirect to login page
   const user = AuthService.getCurrentUser();
@@ -50,22 +83,30 @@ const AddCompany = () => {
     <div className="container-fluid">
       <header className="jumbotron">
         <h3>
-          <span>Add Company</span>
+          <span>Add Contact</span>
         </h3>
       </header>
 
       <FormContainer formContext={formContext} onSuccess={handleSubmitForm}>
         <div>
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
             required
-            name={"name"}
-            label="Name"
+            name={"firstName"}
+            label="First Name"
             variant="outlined"
             margin={"dense"}
           />
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
+            required
+            name={"lastName"}
+            label="Last Name"
+            variant="outlined"
+            margin={"dense"}
+          />
+          <TextFieldElement
+            sx={{ m: 1, width: "52ch" }}
             required
             type={"email"}
             name={"email"}
@@ -73,74 +114,47 @@ const AddCompany = () => {
             margin={"dense"}
             variant="outlined"
           />
+        </div>
 
+        <div>
           <SelectElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
             required
-            options={[
-              {
-                id: "",
-                title: "Select",
-              },
-              {
-                id: "active",
-                title: "active",
-              },
-              {
-                id: "active",
-                title: "Inactive",
-              },
-            ]}
+            options={statusList}
             name={"status"}
             label="Status"
           ></SelectElement>
-        </div>
 
-        <div>
           <SelectElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
             required
-            options={[
-              {
-                id: "",
-                title: "Select",
-              },
-              {
-                id: "customer",
-                title: "Customer",
-              },
-              {
-                id: "carrier",
-                title: "Carrier",
-              },
-            ]}
-            name={"type"}
-            label="Type"
+            options={companies}
+            name={"companyId"}
+            label="Company"
+            labelKey="name"
           ></SelectElement>
 
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
             required
-            name={"revenue"}
-            label="Revenue"
+            name={"department"}
+            label="Department"
             variant="outlined"
-            validation={{ maxLength: 10 }}
-            type={"number"}
-          />
-
-          <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
-            required
-            name={"employeesCount"}
-            label="Employees Count"
-            variant="outlined"
-            validation={{ maxLength: 5 }}
-            type={"number"}
+            validation={{ maxLength: 20 }}
           />
         </div>
         <div>
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
+            required
+            name={"jobTitle"}
+            label="Job Title"
+            variant="outlined"
+            validation={{ maxLength: 50 }}
+          />
+
+          <TextFieldElement
+            sx={{ m: 1, width: "52ch" }}
             required
             name={"address1"}
             label="Address1"
@@ -148,23 +162,22 @@ const AddCompany = () => {
           />
 
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
             name={"address2"}
             label="Address2"
-            variant="outlined"
-          />
-
-          <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
-            required
-            name={"city"}
-            label="City"
             variant="outlined"
           />
         </div>
         <div>
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
+            required
+            name={"city"}
+            label="City"
+            variant="outlined"
+          />
+          <TextFieldElement
+            sx={{ m: 1, width: "52ch" }}
             required
             name={"state"}
             label="State"
@@ -172,25 +185,25 @@ const AddCompany = () => {
           />
 
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
             required
             name={"country"}
             label="Country"
-            variant="outlined"
-          />
-
-          <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
-            required
-            name={"zipcode"}
-            label="Zipcode"
             variant="outlined"
           />
         </div>
 
         <div>
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
+            required
+            name={"zipcode"}
+            label="Zipcode"
+            variant="outlined"
+          />
+
+          <TextFieldElement
+            sx={{ m: 1, width: "52ch" }}
             required
             name={"phone"}
             label="Phone"
@@ -199,7 +212,7 @@ const AddCompany = () => {
           />
 
           <TextFieldElement
-            sx={{ m: 1.1, width: "52ch" }}
+            sx={{ m: 1, width: "52ch" }}
             name={"extension"}
             label="Extension"
             validation={{ maxLength: 6 }}
@@ -226,6 +239,18 @@ const AddCompany = () => {
             >
               Cancel
             </Button>
+
+            {showError != "" ? (
+              <Alert severity="error">{showError}</Alert>
+            ) : (
+              <></>
+            )}
+
+            {showSuccess != "" ? (
+              <Alert severity="success">{showSuccess}</Alert>
+            ) : (
+              <></>
+            )}
           </Stack>
         </div>
       </FormContainer>
@@ -233,4 +258,4 @@ const AddCompany = () => {
   );
 };
 
-export default AddCompany;
+export default AddContact;
