+++ src/components/Home.tsx
@@ -1,3 +1,4 @@
+import React from "react";
 import { Navigate } from "react-router-dom";
 import AuthService from "../services/auth.service";
 const Home = () => {
