--- src/services/auth-header.js
@@ -1,10 +1,0 @@
-export default function authHeader() {
-  const user = JSON.parse(localStorage.getItem("user"));
-
-  if (user && user.token) {
-    return { token: user.token }; // for Spring Boot back-end
-    // return { 'x-access-token': user.accessToken };       // for Node.js Express back-end
-  } else {
-    return {};
-  }
-}
